+++ packages/molecules/header/src/index.tsx
@@ -1,0 +1,2 @@
+import * as React from "react";
+export { Header, type HeaderProps } from "./Header";