+++ packages/theme/src/index.tsx
@@ -2,6 +2,7 @@ export type { Theme } from "./theme";
 export type { Sizes } from "./shared";
 export type { ButtonTheme, ButtonVariant } from "./buttonTheme";
 export type { HeadingTheme } from "./headingTheme";
+export type { HeaderTheme } from "./headerTheme";
 export type { SlideOverTheme } from "./slideOverTheme";
 export {
   useTheme,